Lots of Documentation and Cleanup
