+++ java/simple-cli/src/main/resources/hbase-site.xml
@@ -23,10 +23,10 @@
 -->
 <configuration>
   <property>
-    <name>google.bigtable.project.id</name><value><!-- PROJECT ID --></value>
+    <name>google.bigtable.project.id</name><value>rugged-memory-819</value>
   </property>
   <property>
-    <name>google.bigtable.cluster.name</name><value><!-- CLUSTER NAME --></value>
+    <name>google.bigtable.cluster.name</name><value>cluster</value>
   </property>
   <property>
     <name>google.bigtable.zone.name</name><value>us-central1-b</value>
