+++ java/storm/README.md
@@ -1,7 +1,10 @@
-# Cloud Bigtable Storm Example
+# Cloud Bigtable Storm Example: CoinStorm
 
-This is a sample app that creates a simple Storm topology with a Bolt
-that inserts into Cloud Bigtable.
+This is a sample app that creates a simple [Apache Storm](https://storm.apache.org/) 
+topology with a Spout that reads from the [Coinbase Exchange API WebSocket feed](https://docs.exchange.coinbase.com/#websocket-feed) 
+feed and inserts the data into [Google Cloud Bigtable](cloud.google.com/bigtable).
+
+
 
 ## Provision a Bigtable Cluster
 
@@ -18,12 +21,24 @@ provision the cluster.
 
 ## Download Storm
 
-This example requires you to download a Storm release, both in order to test
-topologies locally as well as to deploy them to remote topologies.
+If you would like to test your topologies locally, download sotrm. 
 
     wget http://www.apache.org/dyn/closer.cgi/storm/apache-storm-0.9.4/apache-storm-0.9.4.tar.gz
     tar -xzf apache-storm-0.9.4.tar.gz
     export PATH=$PATH:$(pwd)/apache-storm-0.9.4/bin
+    
+Note that if you prefer, you can skip testing locally and develop straight in
+ a cluster.
+    
+# Create a Bigtable Table
+
+This examples writes to a table in Bigtable specified on the command line,
+ but assuming it has the column family bc. See the HBase shell examples to see
+ how to create a table with the column family name. The rest of this README
+ assumes you have created a table called Coinbase with a column family "bc". 
+ In the HBase shell, you would do the following:
+ 
+    create 'Coinbase' , { NAME => 'bc' }
 
 ## Set up your hbase-site.xml configuration
 
@@ -100,7 +115,64 @@ and Maven local repository location:
    
 Finally, you can run the topology locally:
 
-    storm jar target/cloud-bigtable-simple-cli-1.0-SNAPSHOT.jar com.example.bigtable.storm.BigtableTopology
+    storm jar target/cloud-bigtable-simple-cli-1.0-SNAPSHOT.jar com.example.bigtable.storm.CoinStormTopology
+    
+## Deploy the Topology
+
+### Create an uberjar
+
+Create a jar with all the necessary dependencies:
+
+    mvn clean package
+
+
+### Download bdutil
+
+bdutil is a tool used to help deploy clusters of GCE instances with common Big
+Data tools installed.
+ 
+[Download bdutil](https://cloud.google.com/hadoop/downloads) or clone it
+on Github [here](https://github.com/GoogleCloudPlatform/bdutil).
+
+### Set up a cluster_config.env
+
+You will need to create a cluster_config.env that specifies some of the 
+configuration of your Storm cluster. Here is a sample cluster_config.sh:
+
+Note that you must create a Google Cloud Storage bucket and specify it in 
+CONFIGBUCKET field. 
+
+    CONFIGBUCKET=your-gcs-bucket
+    PROJECT=your-project-id 
+    PREFIX=your-initials
+    NUM_WORKERS=2
+    GCE_IMAGE='debian-7-backports'
+    GCE_ZONE='us-central1-b'
+
+### Create the Cluster    
+    
+Then deploy the cluster (this might take a few minutes):
+
+    ./bdutil -e cluster_config.sh -e storm
+
+After that you can copy your jar to the master:
+
+    gcloud compute copy-files target/cloud-bigtable-coinstorm-1.0.0.jar your-initials-m:/home/yourusername/
+
+Then ssh into the master:
+    
+    gcloud compute ssh your-initials-m:
+     
+And submit the topology to the Storm cluster (make sure the table Coinbase 
+exists):
+
+    storm jar cloud-bigtable-coinstorm-1.0.0.jar com.example.bigtable.storm.CoinStormTopology Coinbase coinbase_topolog
+
+### View the Topology
+
+Get the external IP of your master instance in the [Cloud Console](console.developer.google.com)
+, then visit <your-ip>:8080 in your browser to view the Storm UI including
+details about your topology.
 
 ## Contributing changes
