+++ java/storm/src/main/java/com/example/bigtable/storm/data/CoinbaseData.java
@@ -1,3 +1,23 @@
+/**
+ * Copyright 2015 Google Inc. All Rights Reserved.
+ * <p/>
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ * <p/>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p/>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.example.bigtable.storm.data;
 
 
@@ -6,9 +26,26 @@ import java.io.Serializable;
 /**
  * This is a POJO object that represents the JSON objects we get from the
  * Coinbase WebSocket streaming API. We use Jackson to deserialize the JSON
- * strings into this object.
+ * strings into this object. Some of these fields are not present in every
+ * instance of incoming data and so will remain as null.
+ *
+ * See here for more info about the Coinbase market feed:
+ * https://docs.exchange.coinbase.com/#websocket-feed
  */
 public class CoinbaseData implements Serializable {
+
+    private String type;
+    private String side;
+    private String order_id;
+    private String remaining_size;
+    private String product_id;
+    private String time;
+    private String sequence;
+    private String price;
+    private String reason;
+    private String size;
+    private String client_oid;
+    private String trade_id;
 
     public String getType() {
         return type;
@@ -105,18 +142,4 @@ public class CoinbaseData implements Serializable {
     public void setTrade_id(String trade_id) {
         this.trade_id = trade_id;
     }
-
-    private String type;
-    private String side;
-    private String order_id;
-    private String remaining_size;
-    private String product_id;
-    private String time;
-    private String sequence;
-    private String price;
-    private String reason;
-    private String size;
-    private String client_oid;
-    private String trade_id;
-
 }
